Rename connection.rs to connection_type.rs
